## Summary
Speed up unit test from 60s to 6s by only mocking Pymarketcap() lib.

Solve the issue: #420

## Quick changelog

- Mock `freqtrade.fiat_convert.Pymarketcap` every time we create a trade
- In `main.create_trade()` move `fiat_converter = CryptoToFiatConverter()` after we saved the the trade into the DB. Today this logic is before we save the trade into the DB. Any failure of Pymarketcap() we fail the saving of the trade in the DB.

## What's new?
Faster unit tests.
Below the new performance of the tests.

|    |   Test  | Original duration | New duration |
|----|---------|-------------------|--------------|
| [x] | freqtrade/tests/test_main.py::test_handle_trade_roi | 6.70s | 0.02s |
| [x] | freqtrade/tests/test_main.py::test_handle_overlpapping_signals | 4.23s | 0.02s |
| [x] | freqtrade/tests/test_main.py::test_sell_profit_only_enable_profit | 4.15s | 0.01s |
| [x] | freqtrade/tests/test_main.py::test_handle_trade_experimental | 4.12s | 0.01s |
| [x] | freqtrade/tests/test_main.py::test_sell_profit_only_disable_loss | 4.10s | 0.01s |
| [x] | freqtrade/tests/test_main.py::test_sell_profit_only_disable_profit | 4.10s | 0.02s |
| [x] | freqtrade/tests/test_main.py::test_process_trade_creation | 2.92s | 0.04s |
| [x] | freqtrade/tests/test_main.py::test_process_trade_handling | 2.36s | 0.02s |
| [x] | freqtrade/tests/test_main.py::test_sell_profit_only_enable_loss | 2.29s | 0.06s |
| [x] | freqtrade/tests/test_main.py::test_execute_sell_without_conf_sell_up | 2.25s | 0.01s |
| [x] | freqtrade/tests/test_main.py::test_close_trade | 2.13s | 0.02s |
| [x] | freqtrade/tests/rpc/test_rpc_telegram.py::test_status_handle | 2.10s | 0.01s |
| [x] | freqtrade/tests/rpc/test_rpc_telegram.py::test_count_handle | 2.09s | 0.01s |
| [x] | freqtrade/tests/rpc/test_rpc_telegram.py::test_status_table_handle | 2.09s | 0.01s |
| [x] | freqtrade/tests/test_main.py::test_execute_sell_without_conf_sell_down | 2.07s | 0.01s |
| [x] | freqtrade/tests/test_main.py::test_create_trade_minimal_amount | 2.06s | 0.01s |
| [x] | freqtrade/tests/rpc/test_rpc_telegram.py::test_performance_handle | 2.03s | 0.02s |
| [x] | freqtrade/tests/test_main.py::test_create_trade | 2.02s | 0.01s |
| [x] | freqtrade/tests/test_fiat_convert.py::test_fiat_convert_is_supported | 1.25s | 0.08s |
| [x] | freqtrade/tests/test_fiat_convert.py::test_fiat_convert_get_price | 1.23s | 0.08s |
| [x] | freqtrade/tests/test_fiat_convert.py::test_fiat_convert_find_price | 1.23s | 0.08s |
| [x] | freqtrade/tests/test_fiat_convert.py::test_fiat_convert_add_pair | 1.19s | 0.08s |