@glonlas 2018-01-21
I am exploring a better way to accelerate unit test by solving the issue at the source.
By changing CryptoToFiatConverter() class into a singleton we will not call Pymarketcap() every time we instance this class. The result will be the same than mocking all call of Pymarketcap() (did in this PR) and will also benefit to the bot itself.