@glonlas 2018-01-22
Another PR will change CryptoToFiatConverter(), the result is the same, except we have not to mock  Pymarketcap() every time we use create_trade()